debugs

+++ app/Jobs/PaysVerification.php
@@ -39,14 +39,13 @@ class PaysVerification implements ShouldQueue
         $pays = Pay::where('verified_by_sistem', NULL)->get();
         $query_params['access_token'] = config('services.mercadopago.token');
         $client = new Client([ 'base_uri' => config('services.mercadopago.base_uri') ]);
-        $delay = now()->subDays(3);
+        $delay = now()->subHour();
+        $four_days = now()->subDays(4);
 
         foreach ($pays as $pay)
         {
-            if(!$pay->collection_id && $pay->created_at < $delay)
-            {
-                $pay->delete();
-            }
+            if(!$pay->collection_id && $pay->created_at < $delay) $pay->delete();
+            if($pay->collection_status = 'Pago creado - Falta verificar' && $pay->created_at < $four_days) $pay->delete();
             if($pay->collection_id)
             {
                $response = $client->request( 'GET', 'v1/payments/' . $pay->collection_id, [
@@ -60,7 +59,7 @@ class PaysVerification implements ShouldQueue
 
                 if($response->status_detail == 'accredited')
                 {
-                    $pay->collection_status = 'Pago recibido'
+                    $pay->collection_status = 'Pago recibido';
                     $pay->verified_by_sistem = now();
                     $pay->update();
                     $device = Device::find($pay->device_id);
